fix: 장바구니 수정 시 옵션 미선택에 관련한 예외처리 추가 (#56)

+++ src/main/java/com/liberty52/product/service/applicationservice/CartItemModifyServiceImpl.java
@@ -60,14 +60,16 @@ public class CartItemModifyServiceImpl implements CartItemModifyService{
       customProduct.modifyCustomPictureUrl(customPictureUrl);
     }
     customProduct.modifyQuantity(dto.getQuantity());
-    customProductOptionRepository.deleteAll(customProduct.getOptions());
-    for (String optionDetailName : dto.getOptions()){
-      CustomProductOption customProductOption = CustomProductOption.create();
-      OptionDetail optionDetail = optionDetailRepository.findByName(optionDetailName)
-          .orElseThrow(() -> new OptionDetailNotFoundException(optionDetailName));
-      customProductOption.associate(optionDetail);
-      customProductOption.associate(customProduct);
-      customProductOptionRepository.save(customProductOption);
+    if (!dto.getOptions().isEmpty()){
+      customProductOptionRepository.deleteAll(customProduct.getOptions());
+      for (String optionDetailName : dto.getOptions()){
+        CustomProductOption customProductOption = CustomProductOption.create();
+        OptionDetail optionDetail = optionDetailRepository.findByName(optionDetailName)
+            .orElseThrow(() -> new OptionDetailNotFoundException(optionDetailName));
+        customProductOption.associate(optionDetail);
+        customProductOption.associate(customProduct);
+        customProductOptionRepository.save(customProductOption);
+      }
     }
   }
